Small fixes
Небольшие правки в дизайне главной страницы, а также переименование некоторых файлов
